fix: pr fixes, small refactoring
