+++ site/plugins/mermaid.ts
